+++ supertokens_python/recipe/thirdparty/providers/apple.py
@@ -22,6 +22,9 @@ from supertokens_python.supertokens import Supertokens
 from jwt import encode, decode
 from time import time
 from re import sub
+from httpx import AsyncClient
+from jwt.algorithms import RSAAlgorithm
+from json import dumps
 
 if TYPE_CHECKING:
     from supertokens_python.framework.request import BaseRequest
@@ -33,6 +36,10 @@ class Apple(Provider):
                  authorisation_redirect: Dict[str, Union[str, Callable[[BaseRequest], str]]] = None,
                  is_default: bool = False):
         super().__init__('apple', client_id, is_default)
+        self.APPLE_PUBLIC_KEY_URL = "https://appleid.apple.com/auth/keys"
+        self.APPLE_PUBLIC_KEY = None
+        self.APPLE_KEY_CACHE_EXP = 60 * 60 * 24
+        self.APPLE_LAST_KEY_FETCH = 0
         default_scopes = ['email']
 
         if scope is None:
@@ -62,6 +69,12 @@ class Apple(Provider):
             r'\\n', '\n', self.client_private_key), algorithm='ES256', headers=headers)
 
     async def get_profile_info(self, auth_code_response: any) -> UserInfo:
+        # - Verify the JWS E256 signature using the server’s public key
+        # - Verify the nonce for the authentication
+        # - Verify that the iss field contains https://appleid.apple.com
+        # - Verify that the aud field is the developer’s client_id
+        # - Verify that the time is earlier than the exp value of the token
+        await self._verify_apple_id_token(auth_code_response['id_token'])
         payload = decode(
             jwt=auth_code_response['id_token'], options={
                 'verify_signature': False})
@@ -99,4 +112,24 @@ class Apple(Provider):
         return AccessTokenAPI(self.access_token_api_url, params)
 
     def get_redirect_uri(self) -> Union[None, str]:
-        return Supertokens.get_instance().app_info.api_domain.get_as_string_dangerous() + APPLE_REDIRECT_HANDLER
+        app_info = Supertokens.get_instance().app_info
+        self.redirect_uri = app_info.api_domain.get_as_string_dangerous()
+        self.redirect_uri += app_info.api_gateway_path.get_as_string_dangerous()
+        self.redirect_uri += app_info.api_base_path.get_as_string_dangerous()
+        self.redirect_uri += APPLE_REDIRECT_HANDLER
+        return self.redirect_uri
+
+    async def _fetch_apple_public_key(self):
+        # Check to see if the public key is unset or is stale before returning
+        if (self.APPLE_LAST_KEY_FETCH + self.APPLE_KEY_CACHE_EXP) < int(time()) or self.APPLE_PUBLIC_KEY is None:
+            async with AsyncClient() as client:
+                response = await client.get(self.APPLE_PUBLIC_KEY_URL)
+                key_payload = response.json()
+                self.APPLE_PUBLIC_KEY = RSAAlgorithm.from_jwk(dumps(key_payload["keys"][0]))
+                self.APPLE_LAST_KEY_FETCH = int(time())
+
+        return self.APPLE_PUBLIC_KEY
+
+    async def _verify_apple_id_token(self, token):
+        public_key = await self._fetch_apple_public_key()
+        decode(token, public_key, audience=self.client_id, algorithm="RS256")