+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When routing, ignores `rid` value `"anti-csrf"`: https://github.com/supertokens/supertokens-python/issues/54
 - `get_redirect_uri` function added to social providers in case we set the `redirect_uri` on the backend.
 - Adds optional `is_default` param to auth providers so that they can be reused with different credentials.
+- Verifies ID Token sent for sign in with apple as per https://developer.apple.com/documentation/sign_in_with_apple/sign_in_with_apple_rest_api/verifying_a_user
 
 ## [0.2.0] - 2021-10-22